COMP: SyncRecord coversion possible loss of data.
  TubeTK\Base\USTK\SyncRecord.cpp(291): warning C4267: 'return' : conversion
  from 'size_t' to 'unsigned int', possible loss of data

+++ Base/USTK/SyncRecord.cpp
@@ -288,7 +288,7 @@ static unsigned int readBinFileAnySizeWithOffset( const char *filename, unsigned
     while( !feof( fp ) ){ totalRead += fread( *bytes, sizeof( unsigned char ), st.st_size-offset, fp ); }
     fclose( fp );
     assert( totalRead == st.st_size-offset );
-    return totalRead;
+    return static_cast< unsigned int >( totalRead );
 }
 //------------------------------------------------------------------------------------------------------------------------------------------
 ///
